requirements.txt
fastapi~=0.115.6
pydantic~=2.10.5
SQLAlchemy~=2.0.37
uvicorn~=0.34.0
pydantic-settings~=2.7.1
python-dotenv~=1.0.1
passlib~=1.7.4